refactor(cache): move invalid route exports to `app/utils/cache.server.ts`
