+++ app/routes/refresh-commit-sha[.]json.tsx
@@ -1,10 +1,10 @@
 import {json} from '@remix-run/node'
-import {cache} from '~/utils/cache.server.ts'
 import {
+  cache,
   isRefreshShaInfo,
   commitShaKey as refreshCacheCommitShaKey,
   type RefreshShaInfo,
-} from './action+/refresh-cache.tsx'
+} from '~/utils/cache.server.ts'
 
 export async function loader() {
   const result = await cache.get(refreshCacheCommitShaKey)
